added ability to initialize BioFVM substrates with initial conditions specified in a .mat file
CSVs (w/o headers) can be used for substrate ics.

Substrate ICs from csv with headers.

Now can include a header row in the CSV defining substrate ICs so that only a subset of substrates can be included for specification. If no header row is given, then it assumes the first n substrates are specified, which could be problematic? It issues a warning about that though.