Relax min value constraint
As upregulation is applied first, it makes absolutely no sense to allow rules to have a max value equal to the base value (it's just wasting our time).
However, downregulation being applied second could make sense for it to equal the base value, having the effect of pulling back down an upregulation event.

This could mean that min values can be omitted and replaced with the base value. But for simplicity, I will keep requiring min values be explicitly supplied when decreasing signals are present.