fix bug in v2 for speed based on anisotropy
I should have been taking the absolute value of cos(theta) when calculating the speed resulting from the fiber alignment. Not doing this meant that moving anti-parallel to a fiber really slowed down movement instead of enhancing movement.

But now I want to use 2 * cos^2(theta) - 1 so that the function is smoother when crossing the perpendicular threshold (so moving from slightly parallel to slightly anti-parallel).