flatten cancer-immune and celltypes3 xmls
these xmls previously relied on a hierarchical structure to define the cell types.
In other words, a "default" cell type would set many of the parameters and then
the individual cell types would override those parameters (I think, though there
may be more nuance to this). This is format for structuring xmls is no
longer in use and (importantly!) is not supported in studio.

this commit flattens all of them. In other words, at the cost of redundancy,
every cell type now has all of its parameters explicitly defined.