fix detachments
- PR #340 put the correct check in the wrong place to prevent detachment when a cell is attacking another cell.
This fixes that by first moving the check into the spring attachments function as an attack produces a spring attachment.
Second, it moves the check into the detachment block rather than the attachment block.

- also, fix bug iterating over the attached cells.
when detaching cells, the for loop would skip checking the last cell in the list if any cells were detached.
So, instead iterate backwards through the list of attached cells.