explicit encoding for pugixml
pugixml will only specify the encoding it it is latin1, i.e. ISO-8859-1.
this might cause problems with some systems that use something called BOM to determine the encoding.
so I have modified the pugixml.cpp to write out the encoding any time it is passed (among cases in a switch statement).